reorganize to move datamodel for Node into the correct file
